+++ lib/flutter_platform_widgets.dart
@@ -7,7 +7,12 @@
 library flutter_platform_widgets;
 
 export 'src/platform.dart'
-    show isCupertino, isMaterial, showPlatformDialog, showPlatformModalSheet;
+    show
+        isCupertino,
+        isMaterial,
+        platform,
+        showPlatformDialog,
+        showPlatformModalSheet;
 export 'src/platform_alert_dialog.dart';
 export 'src/platform_app.dart';
 export 'src/platform_app_bar.dart';