+++ lib/src/platform_icon_button.dart
@@ -8,6 +8,7 @@ import 'package:flutter/cupertino.dart' show CupertinoButton, CupertinoColors;
 import 'package:flutter/material.dart' show IconButton, VisualDensity;
 import 'package:flutter/widgets.dart';
 
+import 'platform.dart';
 import 'widget_base.dart';
 
 const double _kMinInteractiveDimensionCupertino = 44.0;
@@ -100,40 +101,51 @@ class PlatformIconButton extends PlatformWidgetBase<CupertinoButton, Widget> {
   final Key widgetKey;
 
   final Widget icon;
+  @Deprecated('Use cupertinoIcon argument instead')
   final Widget iosIcon;
+  @Deprecated('Use materialIcon argument instead')
   final Widget androidIcon;
+  final Widget cupertinoIcon;
+  final Widget materialIcon;
   final VoidCallback onPressed;
   final Color color;
   final EdgeInsets padding;
   final Color disabledColor;
 
+  @Deprecated('Use material argument. material: (context, platform) {}')
   final PlatformBuilder<MaterialIconButtonData> android;
+  @Deprecated('Use cupertino argument. cupertino: (context, platform) {}')
   final PlatformBuilder<CupertinoIconButtonData> ios;
 
-  PlatformIconButton(
-      {Key key,
-      this.widgetKey,
-      this.icon,
-      this.iosIcon,
-      this.androidIcon,
-      this.onPressed,
-      this.color,
-      this.disabledColor,
-      this.padding,
-      this.android,
-      this.ios})
-      : super(key: key);
+  final PlatformBuilder2<MaterialIconButtonData> material;
+  final PlatformBuilder2<CupertinoIconButtonData> cupertino;
+
+  PlatformIconButton({
+    Key key,
+    this.widgetKey,
+    this.icon,
+    this.iosIcon,
+    this.androidIcon,
+    this.cupertinoIcon,
+    this.materialIcon,
+    this.onPressed,
+    this.color,
+    this.disabledColor,
+    this.padding,
+    this.android,
+    this.ios,
+    this.material,
+    this.cupertino,
+  }) : super(key: key);
 
   @override
-  Widget createAndroidWidget(BuildContext context) {
-    MaterialIconButtonData data;
-    if (android != null) {
-      data = android(context);
-    }
+  Widget createMaterialWidget(BuildContext context) {
+    final data =
+        android?.call(context) ?? material?.call(context, platform(context));
 
     return IconButton(
       key: data?.widgetKey ?? widgetKey,
-      icon: data?.icon ?? androidIcon ?? icon,
+      icon: data?.icon ?? materialIcon ?? androidIcon ?? icon,
       onPressed: data?.onPressed ?? onPressed,
       padding: data?.padding ?? padding ?? const EdgeInsets.all(8.0),
       color: data?.color ?? color,
@@ -154,15 +166,13 @@ class PlatformIconButton extends PlatformWidgetBase<CupertinoButton, Widget> {
   }
 
   @override
-  CupertinoButton createIosWidget(BuildContext context) {
-    CupertinoIconButtonData data;
-    if (ios != null) {
-      data = ios(context);
-    }
+  CupertinoButton createCupertinoWidget(BuildContext context) {
+    final data =
+        ios?.call(context) ?? cupertino?.call(context, platform(context));
 
     return CupertinoButton(
       key: data?.widgetKey ?? widgetKey,
-      child: data?.icon ?? iosIcon ?? icon,
+      child: data?.icon ?? cupertinoIcon ?? iosIcon ?? icon,
       onPressed: data?.onPressed ?? onPressed,
       padding: data?.padding ?? padding,
       color: data?.color ?? color,